qrc files restorer

+++ SpectraSynthesizer.h
@@ -35,7 +35,7 @@ private:
     QJsonObject m_json_config;
     QJsonArray ja;
     QVector<QSlider*> m_sliders;
-    QHash<QString,size_t> lambdas_indexes;
+    QHash<QString,int> lambdas_indexes;
     QSerialPort m_serial_port;
     QSerialPortInfo m_serial_port_info;
     void sendDataToComDevice(const QString command);
